refactor: reorganize packages
