+++ src/main/java/br/ufpr/lpoo/utils/Observer.java
@@ -1,4 +1,4 @@
-package br.ufpr.lpoo.services;
+package br.ufpr.lpoo.utils;
 
 import java.beans.PropertyChangeListener;
